Prepare repo for CMake super project build
When doing a super project build (root CMake file builds shiboken2,
pyside2 and pyside2-tools), there can only be one uninstall target
defined. Make sure not to redefine it if it was already defined.

Also there is no point in doing find_package(Shiboken2) twice,
given that the parent scope of the project already does it.

Change-Id: If8c4cf9e9265632ad29e1dd73d19e6f5cc13ea7f
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>